fix: 修复编译错误
- SequencePropertiesScreen: renderBackground 参数类型 int->float
- PortraitListScreen: NativeImage.read 使用 InputStream 而非 ResourceLocation

+++ src/main/java/top/yourzi/dialog/editor/gui/SequencePropertiesScreen.java
@@ -104,7 +104,7 @@ public class SequencePropertiesScreen extends Screen {
     }
 
     @Override
-    public void renderBackground(GuiGraphics graphics, int mouseX, int mouseY, int partialTick) {
+    public void renderBackground(GuiGraphics graphics, int mouseX, int mouseY, float partialTick) {
         graphics.fill(0, 0, this.width, this.height, 0xFF1A1A1A);
     }
